training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "roboflow>=1.2.11",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "ultralytics>=8.3.239",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv.sources]
"torch" = { index = "pytorch-cu128" }
"torchvision" = { index = "pytorch-cu128" }
"torchaudio" = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true # Optional: use this index only for packages that explicitly specify it
